Add transactionType field to RecordItem
