+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/domain/RecordItem.java
@@ -25,11 +25,14 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import com.hederahashgraph.api.proto.java.Transaction;
 import com.hederahashgraph.api.proto.java.TransactionBody;
 import com.hederahashgraph.api.proto.java.TransactionRecord;
+import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.Value;
+import lombok.extern.log4j.Log4j2;
 
 import com.hedera.mirror.importer.exception.ParserException;
 
+@Log4j2
 @Value
 @AllArgsConstructor
 public class RecordItem implements StreamItem {
@@ -40,6 +43,9 @@ public class RecordItem implements StreamItem {
     private final Transaction transaction;
     private final TransactionBody transactionBody;
     private final TransactionRecord record;
+    // This field is not TransactionTypeEnum since in case of unknown type, we want exact numerical value rather than
+    // -1 in enum.
+    private final int transactionType;
     private final byte[] transactionBytes;
     private final byte[] recordBytes;
 
@@ -66,14 +72,18 @@ public class RecordItem implements StreamItem {
             throw new ParserException(BAD_RECORD_BYTES_MESSAGE, e);
         }
         transactionBody = parseTransactionBody(transaction);
+        transactionType = getTransactionType(transactionBody);
         this.transactionBytes = transactionBytes;
         this.recordBytes = recordBytes;
     }
 
     // Used only in tests
+    // There are many brittle RecordItemParser*Tests which rely on bytes being null. Those tests need to be fixed,
+    // then this function can be removed.
     public RecordItem(Transaction transaction, TransactionRecord record) {
         this.transaction = transaction;
         transactionBody = parseTransactionBody(transaction);
+        transactionType = getTransactionType(transactionBody);
         this.record = record;
         transactionBytes = null;
         recordBytes = null;
@@ -95,5 +105,28 @@ public class RecordItem implements StreamItem {
                 throw new ParserException(BAD_TRANSACTION_BODY_BYTES_MESSAGE, e);
             }
         }
+    }
+
+    /**
+     * Because body.getDataCase() can return null for unknown transaction types, we instead get oneof generically
+     *
+     * @return The protobuf ID that represents the transaction type
+     */
+    private static int getTransactionType(TransactionBody body) {
+        TransactionBody.DataCase dataCase = body.getDataCase();
+
+        if (dataCase == null || dataCase == TransactionBody.DataCase.DATA_NOT_SET) {
+            Set<Integer> unknownFields = body.getUnknownFields().asMap().keySet();
+
+            if (unknownFields.size() != 1) {
+                throw new IllegalStateException("Unable to guess correct transaction type since there's not exactly " +
+                        "one: " + unknownFields);
+            }
+
+            int transactionType = unknownFields.iterator().next();
+            log.warn("Encountered unknown transaction type: {}", transactionType);
+            return transactionType;
+        }
+        return dataCase.getNumber();
     }
 }
